More cleaning in fsm
